fix(playerbot): Fix MSVC compilation - move GroupRole enums to separate header and fix TeleportPlayerToDungeon signature
- Created GroupRoleEnums.h with GroupRole, RoleCapability, and RoleAssignmentStrategy enums
- Updated IRoleAssignment.h to include GroupRoleEnums.h instead of forward declarations
- Updated RoleAssignment.h to include GroupRoleEnums.h instead of defining enums
- Fixed TeleportPlayerToDungeon signature mismatch - added Player* parameter to interface
- Resolves MSVC redefinition errors and virtual function override warnings
